Refactoring Blog Component into Separate Classes
